src/Forecasting/StatsModel/requirements.txt
flask
pandas
numpy
matplotlib
scikit-learn
scipy
google-cloud-storage
flask-cors
cmdstanpy==0.9.5
pystan==2.19.1.1
LunarCalendar>=0.0.9
convertdate>=2.1.2
holidays>=0.10.2
setuptools-git>=1.2
python-dateutil>=2.8.0
tqdm>=4.36.1
pymeeus<=1,>=0.3.13
korean-lunar-calendar
ephem>=3.7.5.3
contourpy>=1.0.1
cycler>=0.10
fonttools>=4.22.0
kiwisolver>=1.0.1
packaging>=20.0
pillow>=6.2.0
pyparsing<3.1,>=2.3.1
tzdata>=2022.1
six>=1.5
statsmodels>=0.12.2
patsy>=0.5.1
pmdarima>=1.8.0
github3.py